Refactor: Gmail Threads 조회 기능 쿼리 조회로 변경
