+++ src/main/java/woozlabs/echo/domain/gmail/service/AsyncGmailService.java
@@ -5,8 +5,8 @@ import com.google.api.services.gmail.model.*;
 import com.google.api.services.gmail.model.Thread;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftListAttachments;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftListDrafts;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftListAttachments;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftListDrafts;
 import woozlabs.echo.domain.gmail.dto.thread.GmailThreadListAttachments;
 import woozlabs.echo.domain.gmail.dto.thread.GmailThreadListThreads;
 import woozlabs.echo.domain.gmail.exception.GmailException;
