+++ src/main/java/woozlabs/echo/domain/gmail/service/GmailService.java
@@ -17,10 +17,10 @@ import org.apache.commons.codec.binary.Base64;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import woozlabs.echo.domain.gmail.dto.*;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftGetMessage;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftGetResponse;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftListDrafts;
-import woozlabs.echo.domain.gmail.dto.darft.GmailDraftListResponse;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftGetMessage;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftGetResponse;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftListDrafts;
+import woozlabs.echo.domain.gmail.dto.draft.GmailDraftListResponse;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageAttachmentResponse;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageSendRequest;
 import woozlabs.echo.domain.gmail.dto.message.GmailMessageSendResponse;
@@ -73,12 +73,12 @@ public class GmailService {
     private final JsonFactory JSON_FACTORY = GsonFactory.getDefaultInstance();
     private final AsyncGmailService asyncGmailService;
 
-    public GmailThreadListResponse getUserEmailThreads(String uid, String pageToken, String category) throws Exception{
+    public GmailThreadListResponse getQueryUserEmailThreads(String uid, String pageToken, String q) throws Exception{
         Member member = memberRepository.findByUid(uid).orElseThrow(
                 () -> new CustomErrorException(ErrorCode.NOT_FOUND_MEMBER_ERROR_MESSAGE));
         String accessToken = member.getAccessToken();
         Gmail gmailService = createGmailService(accessToken);
-        ListThreadsResponse response = getListThreadsResponse(pageToken, category, gmailService);
+        ListThreadsResponse response = getQueryListThreadsResponse(pageToken, q, gmailService);
         List<Thread> threads = response.getThreads(); // get threads
         List<GmailThreadListThreads> detailedThreads = getDetailedThreads(threads, gmailService); // get detailed threads
         Collections.sort(detailedThreads);
@@ -247,13 +247,13 @@ public class GmailService {
         return messages.stream().map(GmailThreadGetMessages::toGmailThreadGetMessages).toList();
     }
 
-    private ListThreadsResponse getListThreadsResponse(String pageToken, String category, Gmail gmailService) throws IOException {
+    private ListThreadsResponse getQueryListThreadsResponse(String pageToken, String q, Gmail gmailService) throws IOException {
         return gmailService.users().threads()
                 .list(USER_ID)
                 .setMaxResults(THREADS_LIST_MAX_LENGTH)
                 .setPageToken(pageToken)
                 .setPrettyPrint(Boolean.TRUE)
-                .setQ(THREADS_LIST_Q + SPACE_CHAR + category)
+                .setQ(q)
                 .execute();
     }
