+++ src/main/java/woozlabs/echo/domain/gmail/util/GmailUtil.java
@@ -1,0 +1,25 @@
+package woozlabs.echo.domain.gmail.util;
+
+import org.springframework.stereotype.Component;
+
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+@Component
+public class GmailUtil {
+    private final Pattern NO_REPLY_PATTERN = Pattern.compile("\\bno.*reply\\b", Pattern.CASE_INSENSITIVE);
+    public Optional<String> extractVerification(String content, String sender){
+        // check sender's email address
+        if(!isVerificationEmail(sender)) return Optional.empty();
+        // Decoding by using base64
+        String code = "test";
+        return Optional.of(code);
+    }
+
+    private boolean isVerificationEmail(String sender){
+        // Additional check on sender's email address
+        Matcher matcher = NO_REPLY_PATTERN.matcher(sender);
+        return matcher.find();
+    }
+}
